pyproject.toml
[tool.poetry]
name = "selfie-sorter"
version = "1.0.0-dev.1"
description = "Local selfie sorter that files images by explicit/suggestive labels using NudeNet + OpenNSFW2"
authors = ["Taylor <t.blackstone@inspyre.tech>"]
readme = "README.md"
packages = [{ include = "selfie_sorter", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pillow = "*"
imagehash = "*"
nudenet = "*"
opennsfw2 = "*"

[tool.poetry.scripts]
selfie-sort = "selfie_sorter.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
docs = [
    "sphinx (>=8.2.3,<9.0.0)",
    "furo (>=2025.9.25,<2026.0.0)",
    "sphinx-autobuild (>=2025.8.25,<2026.0.0)",
    "sphinx-copybutton (>=0.5.2,<0.6.0)",
    "sphinx-design (>=0.6.1,<0.7.0)",
    "myst-parser (>=4.0.1,<5.0.0)"
]
